requirements.txt
Django==1.6.5
South==0.8.4
Unidecode==0.04.16
django-debug-toolbar==1.2.1
django-extensions==1.3.8
djangorestframework==2.3.14
ipython==2.1.0
lxml==3.3.5
python-slugify==0.0.7
requests==2.3.0
six==1.7.2
sqlparse==0.1.11
wsgiref==0.1.2
